ceph: delete measures asynchronously
we should be able to delete measures asynchronously and free up
metricd to do other stuff.

for the unprocessed measures, we first (sync) remove xattr so it
won't be returned in subsequent queries so we don't really care when
object deletion finishes

for stored measures, we first set status to 'delete' in indexer so
it can't be queried so again, we really don't care when object deletion
finishes.

Change-Id: Ib2b5cc7e609c8910bff70f1c794cf007aaa28918