swift: use bulk deletes
we read and write a lot of data constantly. swift supports deleting
objects in bulk[1] so we should leverage this so we can minimise
the number of calls we make to delete objects. this can save
potentially thousands of calls depending on backlog size.

this adds a requirement to configure swift with bulk-filter middleware.

Change-Id: I13afc90d7f67b3d7f213f48f92fc76a1f544861b