carbonara: make sampling mandatory in AggregatedTimeSerie
A long time ago we thought that maybe we would do aggregated timeseries
with no sampling… I'm not sure what it would have make, but whatever,
now it's pretty clear it does not make sense with current code. We don't
use it, but make sure we never do by accident by forbidding it.

Change-Id: I937bef564c77a1bf44d215652b3511ec07940bc1