partition unprocessed measures across workers
this patch assigned each worker to work on a specific partition of
measures backlog. it uses the metric reporting worker to dynamically
set the size of each partition.

by default each processing worker will handle a specific block of 128
metrics. the reporting worker will periodically check backlog size
and broadcast to all processing workers whether to grab larger chunks
(if backlog is big) or smaller chunks (when backlog is small).

this minimises overlap between workers. it may result in no-op
workers if backlog is very small. overlap may occur as the workers
query backlog at different times and may possibly have different block
sizes.

a minimum block size is set at 16 metrics and a maximum of 256.
in theory if 8 workers are set per service, it will grab roughly 128
to 2048 metrics at a time.

Change-Id: Iad7b6b9a56fefcbe85e343e38c0738c2b33efb92