fix timeseries update
related to bug1475329, an error is raised when duplicate timestamps
exist. when we run update, we grab slice of data not in a range.
pandas slicing behaves differently from python slicing where it
includes both start and end endpoints in result[1]. this patch adds a
nanosecond delta so we don't return the endpoints.

http://pandas.pydata.org/pandas-docs/stable/gotchas.html#endpoints-are-inclusive

Change-Id: Ib818acc6928a290a6a297efc6cc372853c17a713
Closes-Bug: #1548448