remove unnecessary sort_index
we don't need to sort_index when using set_values because the ts is
sorted and the ts passed in is sorted (via BoundedTimeSerie). sort_index
is redundant because combine_first maintains ordering when merging.

this patch also uses drop rather than overly complicated double slice
and merging that was happening previously.

Change-Id: Iec84957eceee57a430655379ef4a94a20cd6a6ff
Related-Bug: #1548448