ceph: Try to reduce lock contention
We currently overuses the xattrs on the 'measures' object.
This makes all metricd workers constently waiting to acquire
the object lock to rm_xattr or get_xattrs.

This change is a try to reduce the lock contention on the 'measures' object.

The idea is that all API workers only use 'append' on the 'measures' object.
And 'metricd' workers only use 'append' on a 'measures_done' object.

To list pending measures with just have to do the difference between
this two objects.

Then this two objects need to be regulary cleaned, in this PoC, this is done
while we compute the new set of metrics to be done by a worker.

/!\ It doesn't yet support upgrade process /!\

Closes-bug: #1557724
Change-Id: I0aa5565b08acb0f48d27b86ddfe90610f1f1e0cc