Rename tempest module
This will make clearer the fact that this module should go in
python-gnocchi-tests package for packagers.

We can't move this module in tests otherwise testr/subunit will
try to import them, and that will fail.
We could moves all tests into subdirectory to have

* tests/tempest
* tests/not_tempest

This is a ugly tests reorganisation, so this change only rename
the tempest directory into tests_tempest.

Change-Id: Ic7489421526b6fd9394a2097b397db3945c3dc3e