track resource_type creation/deletion state
The creation and deletion of a resource_type cannot be done
in atomic manner (within a transaction).

We cannot use Sqlalchemy session and metadata.create_all().
And even we can, mysql doesn't support transactional DDL operation.

To workaround this, the change tracks the state of the resource_type
tables creation/deletion, with a state column.

This allows to release the ResourceClassMapper lock during
creation and deletion of a resource type, and reduce the occurrence of
deadlock when too many threads compete to get the sqlalchemy models
of a resource type.

Change-Id: I0ce117b22b50c57bbf5713649eb2af51228e156c