improve task distribution
let a single process gather metrics to process. it will then pipe
the metrics to a queue and all the workers will grab a chunk.

this better distributes tasks to avoid overlap between workers and
also minimises the load on backend where previously all the workers
queried backend for jobs.

Change-Id: I75c9091f09272bccbd859996e4d1745eadf8c329