add support for coordination
use tooz to track all metricd agents. grab a block based on number
of agents active. block size is based on the largest agent with
the largest number of workers

if coordination fails or is not supported, by default, all agents
work against same block with block size based on workers of each
agent.

Change-Id: I291b30f896f5e538040bf8f64d879eb1ad949395