storage: allow to specify regaggregation over aggregation retrieval
Currently, when retrieving e.g. 'mean' aggregation of several metrics,
the aggregation of those metrics is always done using 'mean'.
This patches allows to specify a reaggregation methods, which applies on
top of the measures retrieved.

You can know retrieve e.g. the 'max' of the 'mean' of several metrics.

Related-Bug: #1573023
Change-Id: I9ea60de498e6f81c738a7236a392019b47b0c269