add support to process measures on GET
depending on the backlog, the returned measures from get-measures
may not include what sits in carbonara backlog. to allow for more
flexibility on metricd deployments, this patch offers the ability
to allow users to get all measures even if they don't have enough
metricd workers, at the expense of responsiveness.

Change-Id: I588ae6879474d780e8ec9e893d4ecc2b367b832e
Closes-Bug: #1603495