carbonara: compress non padded timeseries
This creates 2 distinct formats available when serializing
AggregatedTimeSerie: once is "padded" so it can be written based on
offset (Ceph) and the other one is "compressed" so it can be written to
object stores (Swift, file). This compressed format is also used to
store timeseries splits that are out of reach of the back window.

Change-Id: I6f5431945ad9873d280eeb0f6b4907ade493406b