Don't mock ceph
The rados module was mocked because it was not possible to easy ceph
into gate. But now pifpaf support it and we can install anything we want
with bindeps.txt file.

Also future changes will use more and more API from Ceph, making Rados mock class
more hard to maintain.

This change removes the mock and uses a real ceph cluster setuped with
pifpaf to run tests.

Change-Id: I50bd1cac0017565ef678c2ce9235a21368eeea93