Use Cotyledon oslo config glue
Oslo.service was providing some oslo config configuration.
mainly to print all cfg options on startup and a timeout
option in case of service doesn't stop gracefully in an timely fashion.

It also reload the configuration file on sighup.

This change uses the optional cotyledon helper to do that stuffs.

Change-Id: I6227fab55b2f4d33611ca3c4bcf01239ff585748