Fix some gabbi tests
Last gabbi version checks that data have the right content-type.

In theses tests we say that data in plain/text and we put json that's not
correct, old version was silencly ignored and now this is correctly
checked.

Same when we put data but no content-type.

Since our tests are wrong, this change fixes them.

Change-Id: I9a60db296467b08ef32f956471ef5e4f81f462af