Unify timestamp parsing
This unifies all timestamp parsing across Gnocchi. It introduces a new
function to_timestamps() which is responsible for parsing an array of
anything into an array of numpy.datetime64. This array can then be
converted to datetime.datetime object (e.g. for using in the indexer
which supports only microsecond precision) or kept as is for storage
(which supports nanoseconds).

This also include a test plan for making sure we support relative
timestamps even in measures sending (e.g. "-1 minute").

Change-Id: I394a9c14929435504ad6eaa6ce7016fb02f5b26e