Revert "setup: build config file at build time"
Generating the config file at build time is a bit fragile and depends on
how the setup is run (pip, setup.py install, setup.py develop).

For example, pip run bdist_wheel and that fail bdist, both
run the build_py hook. During the bdist_wheel only build dependencies
are available (only pbr for us), so that fail. In our case, it doesn't
because we ignore the exception. Even we block the wheel generation,
bdist will fail too if not all generators dependencies are present.

Perhaps revert this is more secure to ensure we have the configuration
file even it's not up2date.

This reverts commit 17ca48a387f3670d9ac78c4bdebade08c9dba479.

Change-Id: If39c06f8d6101494953813b3fda25f3c7aea9e96