rest: introduce auth_helper to filter resources
This introduce a new plugin system called auth_helper, that can be used
to decide how to filter resources depending on the authentication done.

That does not change the current filtering rules applied with Keystone
authentication, but it removes the artificial filtering that was done
when no authentication method was configured.

Change-Id: I8b6834a10812f16aed808d3a219be9fd86214f4e