try help to Use the pkg_entry info to register the options
I saw this "FIXME(jd) Use the pkg_entry info to register
the options of these libs", and want to try to do it.

I encapsulated the oslo register(and set) methods in some class
in gnocchi opt.py file, so that it allows to use gnocchi entry
point to use them.

Change-Id: I2e632412d04484c746ed776a7c76b1dc490afe41