file: new paths configuration
The current paths configuration doesn't allow
to have a different disk for incoming measurements and
the computed/stored one. And also allow to configure
the tmp drive on an other disk where the code that uses
os.rename() can't ensure that works well.

This change creates instead a new configuration option for
the measure directory and remove the tmp directory configuration.

Instead it creates a tmp directory inside each other directories.

Change-Id: Ice2830633141249dcfc9a11009a44fe6e4ee2279