rest: returns orignal resource id
When batching measures the gnocchi id is returned in 'Unknown Resources' exception.
While the caller have been using non UUID id.

This change returns a dict instead of a list. With the mapping
between the original id and the gnocchi one.

The list to dict convertion allows to keep compatibility with the
the current ceilometer dispatcher.

(see: https://github.com/openstack/ceilometer/blob/master/ceilometer/dispatcher/gnocchi.py#L407)

It can be update later to support both format the list and the dict
and stay compatible with an old and the new Gnocchi exception.

Change-Id: I7b631d17c879b9cd5e56fa8e06b61d5348a182d5