delete unprocessed measures on expunge
instead of cleaning up unprocessed measures of deleted metrics as
they are visible, just defer it to when everything else is cleaned.
this will leave processing workers to only process and cleanup is
done exclusively by janitor.

note: this might schedule delete metrics until cleaned, for now we
assume we can live with this.

Change-Id: Id19d4e1673e660f77576eaee5420d07d7cf65ab2