push incoming into different sacks
having everything in one giant folder/bucket/conatiner/object is bad
because:
- does not allow for good distribution of backend driver.
- makes it hella hard to cleanly split work across multiple metricd
- starves metrics from being processed.

this creates 4096 sacks to distributes them across agents. everything
else pretty much works the same. the agent will loop through all its
sacks based on processing_interval. s3 is set to 4 buckets because it
seems to have a bunch of restrictions.

the number of sacks is constant because it's not safe without more
code which hasn't been agreed upon.

i used sacks to avoid not naming same as driver paradigms.

Change-Id: Icc32d918fe55416385122470c47d60ddbb30dd34