ceph: make sure aio_remove operations return
There's no guarantee that an aio_remove operation will ever finish if Gnocchi
crashes in the middle of an operation. A metric could be entirely expunged from
the indexer while some of its measures/archive files would never be deleted,
leaving stalled files for ever in Ceph.

This patches makes sure this does not happen by waiting for the aio_remove
operation to finish before proceeding to any next steps, or by switching to the
sync remove call.

Change-Id: I7c7bb15a0368c59c14926a1eec226682a9d86a93