hide lock error on delete_metric
tooz non-blocking lock with raise an error if done through
contextmanager. delete_metric uses this functionality. we shouldn't
push an error log in this case since it likely means another process
is handling it. catch error and skip processing, either another
process handles it, or we will come back to it.

Change-Id: I3ab657072706c648cb66eee83a58b69feb919b93