retry to get block on error
When something wrong occurs in coordinator callback we just print a
warning for deployer.But we should also print the root cause and
retry later to get the correct block_size. Otherwise the worker
will never work on its block and an entire will never been processed.

This change tracks the correctness of the block_size and retries if
this one not synced.

Change-Id: Id3fc343cb1556b94af501344a8b016618cbebce3