ceph: store measures in omap values
ceph sucks for small objects; POSTs to ceph incoming are always
small objects.

this changes logic to store new measures as omap values instead of
objects. this allows us to write to 'memory' aka leveldb/rocksdb
instead of disk.

this does not change the durability agreement since we are already
storing object keys in omap so if omap fails, we will lose link
to objects regardless if on disk or not.

testing...

Change-Id: I4bae365955fdbafe4ad837596490774c42bc5251